Planner: extract dml module; add gather_indexes helper
