+++ src/planner/mod.rs
@@ -9,11 +9,13 @@ use schema::resolve_table;
 
 pub(crate) mod resolver;
 pub mod ddl;
+pub(super) mod dml;
+use dml::{plan_insert, plan_update, plan_delete};
 use resolver::{
     ast_expr_name, build_column_mapping, collect_columns, collect_columns_from_column_expr,
-    convert_aggregate, convert_column_expr, convert_expr, convert_having_expr, eval_constant,
+    convert_aggregate, convert_column_expr, convert_expr, convert_having_expr,
     extract_limit_value, extract_table_info, has_aggregate, is_aggregate_function,
-    remap_column_indices, JoinResolver, NoColumnResolver, SingleTableResolver,
+    remap_column_indices, JoinResolver, SingleTableResolver,
 };
 
 // ============================================================================
@@ -345,7 +347,7 @@ pub fn plan(statement: Statement, btree: &BTree) -> Result<LogicalPlan, PlanErro
     }
 }
 
-fn plan_select(select: ast::SelectStatement, btree: &BTree) -> Result<LogicalPlan, PlanError> {
+pub(super) fn plan_select(select: ast::SelectStatement, btree: &BTree) -> Result<LogicalPlan, PlanError> {
     // 1. Extract table info from FROM clause
     let (table_name, table_ref) = extract_table_info(&select.from)?;
 
@@ -882,112 +884,6 @@ fn plan_select_with_joins(
     Ok(plan)
 }
 
-fn output_width(plan: &LogicalPlan) -> usize {
-    match plan {
-        LogicalPlan::Project { columns, .. } => columns.len(),
-        LogicalPlan::Aggregate {
-            group_keys,
-            aggregates,
-            ..
-        } => group_keys.len() + aggregates.len(),
-        LogicalPlan::Count { .. } => 1,
-        LogicalPlan::Values { rows } => rows.first().map(|r| r.len()).unwrap_or(0),
-        LogicalPlan::Sort { input, .. }
-        | LogicalPlan::Limit { input, .. }
-        | LogicalPlan::Distinct { input, .. }
-        | LogicalPlan::Filter { input, .. }
-        | LogicalPlan::RowidLookup { input, .. } => output_width(input),
-        LogicalPlan::Scan { columns, .. } => columns.len(),
-        _ => 0,
-    }
-}
-
-fn plan_insert(insert: ast::InsertStatement, btree: &BTree) -> Result<LogicalPlan, PlanError> {
-    let table = resolve_table(&insert.table_name, btree)?;
-    let num_table_columns = table.columns.len();
-
-    // Determine which columns we're inserting into
-    let table_columns: Vec<usize> = match &insert.columns {
-        Some(col_names) => col_names
-            .iter()
-            .map(|name| {
-                table
-                    .get_column_index(name)
-                    .ok_or_else(|| PlanError::ColumnNotFound {
-                        table: insert.table_name.clone(),
-                        column: name.clone(),
-                    })
-            })
-            .collect::<Result<Vec<_>, _>>()?,
-        None => (0..num_table_columns).collect(),
-    };
-
-    // Build the input plan from the INSERT source
-    let input_plan = match insert.source {
-        ast::InsertSource::Values(value_rows) => {
-            let no_resolver = NoColumnResolver;
-            let mut rows = Vec::new();
-            for value_row in &value_rows {
-                if value_row.len() != table_columns.len() {
-                    return Err(PlanError::ColumnCountMismatch {
-                        expected: table_columns.len(),
-                        got: value_row.len(),
-                    });
-                }
-                let literals: Vec<Literal> = value_row
-                    .iter()
-                    .map(|expr| {
-                        let plan_expr = convert_expr(expr, &no_resolver)?;
-                        eval_constant(&plan_expr)
-                    })
-                    .collect::<Result<Vec<_>, _>>()?;
-                rows.push(literals);
-            }
-            LogicalPlan::Values { rows }
-        }
-        ast::InsertSource::Query(select) => {
-            let input = plan_select(*select, btree)?;
-            let produced = output_width(&input);
-            if produced != table_columns.len() {
-                return Err(PlanError::ColumnCountMismatch {
-                    expected: table_columns.len(),
-                    got: produced,
-                });
-            }
-            input
-        }
-    };
-
-    // Look up indexes for this table
-    let index_infos = btree.lookup_indexes_for_table(&insert.table_name);
-    let mut indexes = Vec::new();
-    for index_info in index_infos {
-        // Find column indexes
-        let column_idxs = index_info
-            .column_names
-            .iter()
-            .map(|name| {
-                table
-                    .columns
-                    .iter()
-                    .position(|col| &col.name == name)
-                    .expect("Index column not found in table")
-            })
-            .collect();
-
-        indexes.push(IndexMaintenanceInfo {
-            rootpage: index_info.rootpage,
-            column_idxs,
-        });
-    }
-
-    Ok(LogicalPlan::Insert {
-        rootpage: table.rootpage,
-        table_columns,
-        input: Box::new(input_plan),
-        indexes,
-    })
-}
 
 /// Try to use an index for the given filter (equality or range predicate).
 fn try_plan_index_scan(
@@ -1167,113 +1063,6 @@ fn extract_literal(expr: &ast::Expression) -> Option<Literal> {
         _ => None,
     }
 }
-
-fn plan_update(update: ast::UpdateStatement, btree: &BTree) -> Result<LogicalPlan, PlanError> {
-    let table = resolve_table(&update.table_name, btree)?;
-
-    // Build column mapping: all table columns in order
-    let column_map = table.column_name_map();
-
-    // Create column resolver
-    let resolver = SingleTableResolver {
-        table_ref: &update.table_name,
-        columns: &column_map,
-    };
-
-    // Resolve assignment column names to indices and plan their expressions
-    let mut assignments = Vec::new();
-    for (col_name, expr) in update.assignments {
-        let col_idx = table
-            .columns
-            .iter()
-            .position(|c| c.name == col_name)
-            .ok_or_else(|| PlanError::ColumnNotFound {
-                table: update.table_name.clone(),
-                column: col_name.clone(),
-            })?;
-
-        // Plan the expression (column refs refer to table schema)
-        let expr_plan = convert_expr(&expr, &resolver)?;
-        assignments.push((col_idx, expr_plan));
-    }
-
-    // Plan the filter expression if present
-    let filter = match update.filter {
-        Some(expr) => Some(convert_expr(&expr, &resolver)?),
-        None => None,
-    };
-
-    // Get all table column indices
-    let table_columns: Vec<usize> = (0..table.columns.len()).collect();
-
-    // Gather secondary index maintenance info (mirrors plan_delete)
-    let index_infos = btree.lookup_indexes_for_table(&update.table_name);
-    let mut indexes = Vec::new();
-    for index_info in index_infos {
-        let column_idxs = index_info
-            .column_names
-            .iter()
-            .map(|name| table.columns.iter().position(|c| &c.name == name).unwrap())
-            .collect();
-        indexes.push(IndexMaintenanceInfo {
-            rootpage: index_info.rootpage,
-            column_idxs,
-        });
-    }
-
-    Ok(LogicalPlan::Update {
-        rootpage: table.rootpage,
-        table_columns,
-        assignments,
-        filter,
-        indexes,
-    })
-}
-
-fn plan_delete(delete: ast::DeleteStatement, btree: &BTree) -> Result<LogicalPlan, PlanError> {
-    let table = resolve_table(&delete.table_name, btree)?;
-
-    // Build column mapping: all table columns in order
-    let column_map = table.column_name_map();
-
-    // Create column resolver
-    let resolver = SingleTableResolver {
-        table_ref: &delete.table_name,
-        columns: &column_map,
-    };
-
-    // Plan the filter expression if present
-    let filter = match delete.filter {
-        Some(expr) => Some(convert_expr(&expr, &resolver)?),
-        None => None,
-    };
-
-    // Get all table column indices
-    let table_columns: Vec<usize> = (0..table.columns.len()).collect();
-
-    // Gather secondary index maintenance info (mirrors plan_insert)
-    let index_infos = btree.lookup_indexes_for_table(&delete.table_name);
-    let mut indexes = Vec::new();
-    for index_info in index_infos {
-        let column_idxs = index_info
-            .column_names
-            .iter()
-            .map(|name| table.columns.iter().position(|c| &c.name == name).unwrap())
-            .collect();
-        indexes.push(IndexMaintenanceInfo {
-            rootpage: index_info.rootpage,
-            column_idxs,
-        });
-    }
-
-    Ok(LogicalPlan::Delete {
-        rootpage: table.rootpage,
-        table_columns,
-        filter,
-        indexes,
-    })
-}
-
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum PlanError {
